Заполнены почти все поля

+++ shop/main/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from django.urls import reverse
 
+
 class Category(models.Model):
     name = models.CharField(max_length=100, db_index=True)
     slug = models.SlugField(max_length=100, unique=True)
@@ -15,15 +16,15 @@ class Category(models.Model):
 
     def get_absolute_url(self):
         return reverse("main:product_list_by_category", args=[self.slug])
-    
+
 
 class Product(models.Model):
     category = models.ForeignKey(Category, related_name='product', on_delete=models.CASCADE)
     name = models.CharField(max_length=100, db_index=True)
     slug = models.SlugField(max_length=100, unique=True)
-    image = models.ImageField(upload_to='products/%Y/%m/%d' )
+    image = models.ImageField(upload_to='products/%Y/%m/%d')
     description = models.TextField(blank=True)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.IntegerField()
     available = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
     updated = models.DateTimeField(auto_now=True)
@@ -35,6 +36,6 @@ class Product(models.Model):
 
     def __str__(self):
         return self.name
+
     def get_absolute_url(self):
-        return reverse("main:product_detail", args=[self.id,self.slug])
-    
+        return reverse("main:product_detail", args=[self.id, self.slug])
